[learn-squeryl] id is now an 'int'.

+++ learn_scala/squeryl/src/main/scala/pkg/db/Custom.scala
@@ -1,4 +1,4 @@
 package pkg.db
 
 // Uses a custom type
-case class Custom(id: Long, cust: RevString) extends HasId
+case class Custom(id: Int, cust: RevString) extends HasId
